Move interaction test views to submodules
